requirements.txt
streamlit
langchain>=0.2.10
langchain-community>=0.2.10
langchain-core>=0.2.10
langchain-text-splitters>=0.2.0
langchain-google-genai>=1.0.7
langchain-openai>=0.1.22
langchain-huggingface
faiss-cpu
sentence-transformers
python-dotenv
google-generativeai
pypdf
pyarrow-hotfix